+++ ldm_core/constants.py
@@ -1,8 +1,8 @@
 from pathlib import Path
 
 # --- Constants & Configuration ---
-# LDM_MAGIC_VERSION: 2.7.2-beta.19
-VERSION = "2.7.2-beta.19"
+# LDM_MAGIC_VERSION: 2.7.2-beta.22
+VERSION = "2.7.2-beta.22"
 
 
 BUILD_INFO = None